+++ extension/src/LanguageFunctions.ts
@@ -17,6 +17,7 @@ import { LanguageFunctionsSettings } from "./Settings/LanguageFunctionsSettings"
 import * as XliffFunctions from "./XliffFunctions";
 import { XliffIdToken } from "./ALObject/XliffIdToken";
 import { TextDocumentMatch } from "./Types";
+import { logger } from "./Logging/LogHelper";
 
 export async function findNextUntranslatedText(
   filesToSearch: string[],
@@ -83,7 +84,7 @@ export async function findNextUntranslatedText(
     if (!xlfDocument.translationTokensExists()) {
       if (xlfDocument.customNotesOfTypeExists(CustomNoteType.refreshXlfHint)) {
         xlfDocument.removeAllCustomNotesOfType(CustomNoteType.refreshXlfHint);
-        console.log("Removed custom notes.");
+        logger.log("Removed custom notes.");
         xlfDocument.toFileAsync(xlfPath, replaceSelfClosingXlfTags);
       }
     }
@@ -361,10 +362,10 @@ export function getTransUnitID(
   return { lineNo: activeLineNo - count + 1, id: result[1] };
 }
 
-export async function revealTransUnitTarget(
+export function revealTransUnitTarget(
   transUnitId: string,
   langFilePath: string
-): Promise<TextDocumentMatch | undefined> {
+): TextDocumentMatch | undefined {
   if (!vscode.window.activeTextEditor) {
     return;
   }