+++ extension/src/Settings/Settings.ts
@@ -55,6 +55,8 @@ export class Settings {
   public xliffCSVImportTargetState = "translated";
   public loadSymbols = true;
   public refreshXlfAfterFindNextUntranslated = false;
+  public enableTranslationsOnHover = true;
+  public enableTelemetry = true;
   public useDictionaryInDTSImport = true;
 
   constructor(workspaceFolderPath: string) {