+++ extension/src/Settings/SettingsMap.ts
@@ -50,4 +50,6 @@ export const settingsMap = new Map<string, keyof Settings>([
     "NAB.RefreshXlfAfterFindNextUntranslated",
     "refreshXlfAfterFindNextUntranslated",
   ],
+  ["NAB.EnableTranslationsOnHover", "enableTranslationsOnHover"],
+  ["NAB.EnableTelemetry", "enableTelemetry"],
 ]);